refactor: extract algorithm resources
